requirements.txt
numpy==1.19.5
matplotlib==3.3.3
attrs==20.3.0
beautifulsoup4==4.9.3
requests==2.25.1
selenium==3.141.0
stem==1.8.0
pandas==1.2.0
torpy==1.1.5
